+++ src/pages/ModifyProfile/ModifyPassword/style.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const Form = styled.form`
+  width: 50%;
+  min-width: 300px;
+  max-width: 400px;
+  height: auto;
+  padding: 30px 0;
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+`;
